+++ crates/arkworks/src/bn254.rs
@@ -14,33 +14,6 @@ impl CurveBasicOperations for Bn254 {
     const G1_LEN: usize = 65;
     const G2_LEN: usize = 129;
     const SCALAR_LEN: usize = 32;
-    const CURVE_ID: usize = 0x2c;
 }
 
-#[test]
-fn test_wasm_pairing() {
-    let mut rng = test_rng();
-    let a: G1Projective = rng.gen();
-    let b: G2Projective = rng.gen();
-    let s: Fr = rng.gen();
-
-    let mut sa = a;
-    sa.mul_assign(s);
-    let mut sb = b;
-    sb.mul_assign(s);
-
-    let ans1 = <Bn254 as PairingEngine>::pairing(sa, b);
-    let ans2 = <Bn254 as PairingEngine>::pairing(a, sb);
-    let ans3 = <Bn254 as PairingEngine>::pairing(a, b).pow(s.into_repr());
-
-    assert_eq!(ans1, ans2);
-    assert_eq!(ans2, ans3);
-
-    assert_ne!(ans1, Fq12::one());
-    assert_ne!(ans2, Fq12::one());
-    assert_ne!(ans3, Fq12::one());
-
-    assert_eq!(ans1.pow(Fr::characteristic()), Fq12::one());
-    assert_eq!(ans2.pow(Fr::characteristic()), Fq12::one());
-    assert_eq!(ans3.pow(Fr::characteristic()), Fq12::one());
-}
+paste_test!();
