+++ crates/arkworks/src/cp6_782.rs
@@ -12,33 +12,6 @@ impl CurveBasicOperations for CP6_782 {
     const G1_LEN: usize = 209;
     const G2_LEN: usize = 625;
     const SCALAR_LEN: usize = 48;
-    const CURVE_ID: usize = 0x2d;
 }
 
-#[test]
-fn test_wasm_pairing() {
-    let mut rng = test_rng();
-    let a: G1Projective = rng.gen();
-    let b: G2Projective = rng.gen();
-    let s: Fr = rng.gen();
-
-    let mut sa = a;
-    sa.mul_assign(s);
-    let mut sb = b;
-    sb.mul_assign(s);
-
-    let ans1 = <CP6_782 as PairingEngine>::pairing(sa, b);
-    let ans2 = <CP6_782 as PairingEngine>::pairing(a, sb);
-    let ans3 = <CP6_782 as PairingEngine>::pairing(a, b).pow(s.into_repr());
-
-    assert_eq!(ans1, ans2);
-    assert_eq!(ans2, ans3);
-
-    assert_ne!(ans1, Fq6::one());
-    assert_ne!(ans2, Fq6::one());
-    assert_ne!(ans3, Fq6::one());
-
-    assert_eq!(ans1.pow(Fr::characteristic()), Fq6::one());
-    assert_eq!(ans2.pow(Fr::characteristic()), Fq6::one());
-    assert_eq!(ans3.pow(Fr::characteristic()), Fq6::one());
-}
+paste_test!();
